+++ src/classes/providers/completionitems/factories/xml/StandardXMLCompletionItemFactory.ts
@@ -112,7 +112,7 @@ export class StandardXMLCompletionItemFactory implements ICompletionItemFactory 
 		const className = data.className.split(".")[data.className.split(".").length - 1];
 		const completionItem: CustomCompletionItem = new CustomCompletionItem(className);
 		completionItem.kind = vscode.CompletionItemKind.Class;
-		completionItem.insertText = data.insertText;
+		completionItem.insertText = new vscode.SnippetString(data.insertText);
 		completionItem.className = data.className;
 		completionItem.detail = data.detail;
 		completionItem.documentation = data.markdown;
@@ -151,9 +151,20 @@ export class StandardXMLCompletionItemFactory implements ICompletionItemFactory 
 
 	private _generateInsertStringFrom(className: string, properties: string, aggregations: string, classPrefix: string, prefixBeforeClassName = "") {
 		let insertText = `${prefixBeforeClassName}${className}\n`;
-		insertText += properties;
+
+		const shouldAttributesBeAdded = vscode.workspace.getConfiguration("ui5.plugin").get("addTagAttributes");
+		const shouldAggregationsBeAdded = vscode.workspace.getConfiguration("ui5.plugin").get("addTagAggregations");
+
+		if (shouldAttributesBeAdded) {
+			insertText += properties;
+		}
 		insertText += ">\n";
-		insertText += aggregations;
+
+		if (shouldAggregationsBeAdded) {
+			insertText += aggregations;
+		} else {
+			insertText += "    $0\n";
+		}
 
 		insertText += `</${classPrefix}${className}>`;